fix(!go): fix handling of '!go http'
'!go http' needs to redirect to pkg.go.dev/net/http, but this does not happen
because we only redirect if the full package name (net/http) is specified.

Make the search fuzzier by allowing matches to the import name. This allows
'!go http' and '!go url.QueryEscape' to jump straight to the package docs.